+++ marimbabot_command_recognition/package.xml
@@ -14,7 +14,6 @@
   <buildtool_depend>catkin</buildtool_depend>
   <build_depend>rospy</build_depend>
   <build_depend>std_msgs</build_depend>
-  <build_depend>sound_play</build_depend>
   <build_depend>message_generation</build_depend>
 
   <build_export_depend>rospy</build_export_depend>
@@ -23,5 +22,4 @@
   <exec_depend>rospy</exec_depend>
   <exec_depend>std_msgs</exec_depend>
   <exec_depend>message_runtime</exec_depend>
-  <exec_depend>sound_play</exec_depend>
 </package>
