+++ marimbabot_command_recognition/srv/Speak.srv
@@ -1,0 +1,4 @@
+int64 id
+string text
+---
+bool is_end
